+++ spec/support/webmock_server.rb
@@ -31,8 +31,8 @@ class WebMockServer
         trap(signal){ server.shutdown }
       end
       server.start do |socket|
-        socket.puts <<-EOT.gsub(/^\s+\|/, '')
-          |HTTP/1.0 200 OK
+        socket.puts <<-EOT.gsub(/^\s+\|/, '').gsub("\n","\r\n")
+          |HTTP/1.1 200 OK
           |Date: Fri, 31 Dec 1999 23:59:59 GMT
           |Content-Type: text/html
           |Content-Length: 11